ClippingAttachmentModel updated
