+++ Sources/Spine/Model/Skin/Attachments/ClippingAttachmentModel.swift
@@ -1,0 +1,38 @@
+//
+//  ClippingAttachmentModel.swift
+//  
+//
+//  Created by Max Gribov on 07.11.2022.
+//
+
+import SpriteKit
+
+struct ClippingAttachmentModel: AttachmentModel {
+    
+    //TODO: - remove
+    let name = ""
+    
+    let end: String
+    let vertexCount: UInt
+    let vertices: [CGFloat]
+    let color: ColorModel
+}
+
+extension ClippingAttachmentModel: SpineDecodableDictionary {
+    
+    enum Keys: String, CodingKey {
+        
+        case name, end, vertexCount, vertices, color
+    }
+    
+    typealias KeysType = Keys
+    
+    init(_ name: String, _ container: KeyedDecodingContainer<KeysType>) throws {
+        
+        end = try container.decode(String.self, forKey: .end)
+        vertexCount = try container.decode(UInt.self, forKey: .vertexCount)
+        vertices = try container.decode([CGFloat].self, forKey: .vertices)
+        let colorValue = try container.decodeIfPresent(String.self, forKey: .color) ?? "CE3A3AFF"
+        color = ColorModel(colorValue)
+    }
+}
